feat #2 : 기초 Controller, Dto, service, mapper, entity, repository 작성

+++ server/src/main/java/codestates/main22/tree/dto/TreeDto.java
@@ -17,7 +17,7 @@ public class TreeDto {
         private int treePoint;
         private String treeImage;
 
-        public void setTreeId(int treeId) {this.treeId = treeId;}
+        public void setTreeId(long treeId) {this.treeId = treeId;}
     }
 
     @Getter
